% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdiscpilot01_adsl}
\alias{cdiscpilot01_adsl}
\title{CDISCPILOT01 subject-level analysis dataset (ADSL)}
\format{
A tibble, one row per subject.
}
\source{
CDISC SDTM/ADaM pilot project (CDISCPILOT01).
}
\usage{
cdiscpilot01_adsl
}
\description{
The ADaM ADSL for the CDISC pilot study, bundled as sample data for the
execution examples.
}
\keyword{datasets}
