% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_importModule.R
\name{check_signature_validity}
\alias{check_signature_validity}
\title{Check validity of signature}
\usage{
check_signature_validity(graph_list, public_key)
}
\arguments{
\item{graph_list}{list containing graph elements}

\item{public_key}{public key}
}
\description{
Check validity of signature
}
