% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_library.R
\name{cl_products}
\alias{cl_products}
\title{List CDISC Library products}
\usage{
cl_products(key = cdisc_library_key())
}
\arguments{
\item{key}{A CDISC Library API key; defaults to \code{\link[=cdisc_library_key]{cdisc_library_key()}}.}
}
\value{
A tibble of product hrefs discovered under \verb{/mdr/products}.
}
\description{
List CDISC Library products
}
