% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{copy_trace}
\alias{copy_trace}
\title{Copy an item of tracer object}
\usage{
copy_trace(x, what)
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{either numeric indexes or a vector string of UID. 
If some numeric indexes out of range an error will be thrown. 
All Unmatched UID's will be ignored with warning.}
}
\value{
object of type tracer with selected traces.
}
\description{
copy selected traces from the tracer object
}
