% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageReport.R
\name{Coverage}
\alias{Coverage}
\alias{summariseTracesAndTrees}
\title{Coverage Report from Model Validation}
\usage{
summariseTracesAndTrees(log.file, tree.file = NA, burn.in = 0.1)
}
\arguments{
\item{log.file}{The BEAST log file containing posterior samples.}

\item{tree.file}{The BEAST tree log file containing posterior trees.
Default to NA, which will skip to report the tree statistics.}

\item{burn.in}{The proportion of samples treated as the burn-in stage
during MCMC. Default to 0.1.}
}
\description{
The coverage report of a model validation result.
}
\details{
\code{summariseTracesAndTrees} creates the statistical summary of
traces from a BEAST log file and tree statistics from BEAST tree log file,
and save them into tsv files.
}
\examples{
WD = file.path("~/WorkSpace/MyValidations/")
setwd(WD)
log.files = list.files(pattern = ".log")
for(lg in log.files) {
  tree.file=paste0(sub('\\\\.log$', '', lg), ".trees")
  summariseTracesAndTrees(lg, tree.file)
}

}
\keyword{Coverage}
