% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_downloadModule.R
\name{createDownloadModuleUI}
\alias{createDownloadModuleUI}
\title{Dynamically add ui elements for downloadButton}
\usage{
createDownloadModuleUI(output)
}
\arguments{
\item{output}{shiny session output object}
}
\description{
Dynamically add ui elements for downloadButton
}
