% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_importHelper.R
\name{createGraphFromUpload}
\alias{createGraphFromUpload}
\title{Creates a graph from a (uploaded) list}
\usage{
createGraphFromUpload(graph_list)
}
\arguments{
\item{graph_list}{list, containing all elements required for the graph}
}
\description{
Creates a graph from a (uploaded) list
}
