% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_point_validator}
\alias{data_point_validator}
\title{Prompt for re-sampling
depends on `trace_info()`, `tr_resample()`}
\usage{
data_point_validator(x)
}
\description{
Prompt for re-sampling
depends on `trace_info()`, `tr_resample()`
}
\keyword{internal}
