% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_displayNodeID.R
\name{displayNodeId}
\alias{displayNodeId}
\title{Displays the node Id that was clicked}
\usage{
displayNodeId(input, output, outputId)
}
\arguments{
\item{input}{server input object}

\item{output}{server output object}

\item{outputId}{outputId of element for which the clicked Id is shown}
}
\description{
Displays the node Id that was clicked
}
