% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_list.R
\name{exposure_list}
\alias{exposure_list}
\title{Extract individual exposures from an epicontacts outbreak}
\usage{
exposure_list(outbreak, date_onset_column)
}
\arguments{
\item{outbreak}{the outbreak as an epicontacts object}

\item{date_onset_column}{the column name of for the date of onset}
}
\value{
a list with one entry per individual. Each entry is a vector
        of exposure dates
}
\description{
Extract individual exposures from an epicontacts outbreak
}

