% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_scan}
\alias{file_scan}
\title{a Parser helper}
\usage{
file_scan(fls, delim = NULL)
}
\description{
Tries different delimiters and determines a file Template to pass to
the parser selector
}
\keyword{internal}
