% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{gen_uid}
\alias{gen_uid}
\title{Generates unique ID}
\usage{
gen_uid(len)
}
\arguments{
\item{len}{the length or the number of symbols in UID}
}
\description{
An ID generator that produces a single ID code composed of
lower case letters and 10 digits with prefix 'w'
}
