% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{get_meta_fields}
\alias{get_meta_fields}
\title{Get meta data field description}
\usage{
get_meta_fields(x = NULL)
}
\arguments{
\item{x}{objects of class tracer. If an object of class tracer is not supplied, 
a default list of fields and description will be returned.}
}
\description{
returns fields of mete data across all traces
}
