% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_importModule.R
\name{importModuleServer}
\alias{importModuleServer}
\title{Import JSON Module Server}
\usage{
importModuleServer(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
uploaded data
}
\description{
Import JSON Module Server
}
