% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_importModule.R
\name{importModuleUI}
\alias{importModuleUI}
\title{Import JSON Module UI}
\usage{
importModuleUI(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI elements
}
\description{
Import JSON Module UI
}
