% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{init_log}
\alias{init_log}
\title{Keeps track of processing steps taken and records them into a data.frame}
\usage{
init_log(what = NULL, tmpl = c("LOG", "META"))
}
\arguments{
\item{what}{either NULL or a list with named fields and values}
}
\description{
Keeps track of processing steps taken and records them into a data.frame
}
\keyword{internal}
