% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inspect_trace}
\alias{inspect_trace}
\title{Inspect trace}
\usage{
inspect_trace(data)
}
\arguments{
\item{data}{Data frame of traced coordinates i.e. the output of `trace_image`. Must have values
`x` and `y`}
}
\value{
ggplot image
}
\description{
Takes as input the traced coordinates and plots them as a polygon.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggimage)

# plot an image
img <- paste0(system.file(package = "traceR"), "/images/star.png")
ggplot() +
  geom_image(aes(0, 0, image = img), size = 1)

# trace the image
df <- trace_image()

# inspect the traced image
inspect_trace(df)
\dontshow{\}) # examplesIf}
}
