% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge.R
\name{is_trace_graph}
\alias{is_trace_graph}
\title{Test whether an object is a trace graph}
\usage{
is_trace_graph(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} is a \code{\link[=trace_graph]{trace_graph()}}.
}
\description{
Test whether an object is a trace graph
}
