% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Logger}
\alias{Logger}
\title{Simple logging utility}
\description{
Creates an Logger class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logger-new}{\code{Logger$new()}}
\item \href{#method-Logger-set_name}{\code{Logger$set_name()}}
\item \href{#method-Logger-set_log_level}{\code{Logger$set_log_level()}}
\item \href{#method-Logger-set_printer}{\code{Logger$set_printer()}}
\item \href{#method-Logger-trace}{\code{Logger$trace()}}
\item \href{#method-Logger-debug}{\code{Logger$debug()}}
\item \href{#method-Logger-info}{\code{Logger$info()}}
\item \href{#method-Logger-warn}{\code{Logger$warn()}}
\item \href{#method-Logger-error}{\code{Logger$error()}}
\item \href{#method-Logger-fatal}{\code{Logger$fatal()}}
\item \href{#method-Logger-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(
  level = c("info", "fatal", "error", "warn", "debug", "trace", "off", "all"),
  name = "logger",
  printer = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$set_name(name = "logger")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-set_log_level"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-set_log_level}{}}}
\subsection{Method \code{set_log_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$set_log_level(
  level = c("info", "fatal", "error", "warn", "debug", "trace", "off", "all")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-set_printer"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-set_printer}{}}}
\subsection{Method \code{set_printer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$set_printer(FUN = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-trace"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-trace}{}}}
\subsection{Method \code{trace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$trace(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-debug"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-debug}{}}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$debug(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-info"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$info(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-warn"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$warn(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-error"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-error}{}}}
\subsection{Method \code{error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$error(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-fatal"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-fatal}{}}}
\subsection{Method \code{fatal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$fatal(msg, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
