% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{merge_trace}
\alias{merge_trace}
\title{Merge objects}
\usage{
merge_trace(a, b, what = NULL, active_re = TRUE, keep_history = FALSE)
}
\arguments{
\item{a, b}{objects of class tracer}

\item{what}{an optional argument either numeric indexes or vector string of UID, 
that allows selecting traces from object `b` to be merged into object `a`. Default is NULL.}

\item{active_re}{logical. If TRUE recomputes UID's to avoid duplicates.}

\item{keep_history}{logical, whether processing history of object `b` should be kept 
in the new object or not.}
}
\value{
object of type tracer.
}
\description{
merges two trace objects into one
}
\details{
If  argument `active_re` is FALSE than all identical UIDs found 
in objects `a` and `b` will be ignored and not merged into object `a`.
In case `a` and `b` are identical an object `a` will be returned.
}
