% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperatorProposal.R
\name{Operator}
\alias{Operator}
\alias{readState}
\title{Summarising proposals for the operator}
\usage{
readState(file, acc.cols = c("accept", "reject"), id.cols = c("id", "p"))
}
\arguments{
\item{file}{The file to read/write.}

\item{acc.cols, id.cols}{The column names which matches the names in json.}
}
\description{
Extract the sumary of operator proposals from BEAST 2 \url{http://www.beast2.org} output.

Tuning: The value of the operator's tuning parameter, or '-' if the operator can't be optimized.
accept: The total number of times a proposal by this operator has been accepted.
reject: The total number of times a proposal by this operator has been rejected.
Pr(m): The probability this operator is chosen in a step of the MCMC (i.e. the normalized weight).
Pr(acc|m): The acceptance probability (\code{accept} as a fraction of the total proposals for this operator).
}
\details{
\code{readState} reads MCMC state log file and return a data frame
to summarise operators' acceptence and rejection.
}
\keyword{Operator}
