% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pal1}
\alias{pal1}
\alias{palettes}
\title{Color palettes used in tracer}
\usage{
pal1(n)
}
\arguments{
\item{n}{a number of colors}
}
\description{
These functions are color palettes used in tracer.
}
\details{
\itemize{
\item pal1: purples (dark to light)
}
}
\examples{

plot(1:100, cex=8, pch=20, col = pal1(100),
     main="palette: pal1")

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

