% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_file}
\alias{parse_file}
\title{A simple parser , imports .csv, .txt, .arw files}
\usage{
parse_file(fls, read_par)
}
\description{
A simple parser , imports .csv, .txt, .arw files
}
\keyword{internal}
