% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{plot.tracer}
\alias{plot.tracer}
\title{Plot traces}
\usage{
\method{plot}{tracer}(x, ...)
}
\arguments{
\item{x}{an object of class tracer}

\item{...}{additional arguments to pass to plt_gg, see ?plt_gg}
}
\description{
this function is - under development -
currently 'ggplot2' is only supported graphic system. 'base' and 'plotly' are coming next.
}
