% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{+.tracer}
\alias{+.tracer}
\title{Operator `+` overloading}
\usage{
\method{+}{tracer}(a, b)
}
\arguments{
\item{a, b}{objects of class tracer}
}
\description{
Operator `+` overloading
}
\details{
a wrapper for `merge_trace`, where keep_history and re_active are TRUE.
}
