% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rehash_id}
\alias{rehash_id}
\title{Rehashing indexes - a helper for merging objects}
\usage{
rehash_id(obj, pool = NULL)
}
\description{
Rehashing indexes - a helper for merging objects
}
\keyword{internal}
