% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_renderFlowchart.R
\name{renderFlowchart}
\alias{renderFlowchart}
\title{Render flowchart}
\usage{
renderFlowchart(graph)
}
\arguments{
\item{graph}{reactive graph object}
}
\value{
rendered flowchart
}
\description{
Render flowchart
}
