% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_coords}
\alias{scale_coords}
\title{Scale coords}
\usage{
scale_coords(x)
}
\arguments{
\item{x}{Vector of values to be scaled}
}
\value{
Vector of numeric values
}
\description{
Scales recorded coordinates to (0, 1)
}
\examples{
x <- 1:10
scale_coords(x)
}
