% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_field}
\alias{set_field}
\title{Set a new records to a selected field of the meta data}
\usage{
set_field(x, what, field, full_match = FALSE)
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{a smart row selector see `plt_gg` for details.}

\item{field}{a named list of fields and corresponding records}

\item{full_match}{logical, if set TRUE a field name if exists will be searched 
for exact match.}
}
\value{
an object of type tracer with new fields and records updated
}
\description{
adding new meta data to the selected samples
}
\details{
depending on what is passed to `what`, meta data update can be 
done to a single sample or group of samples. If `what` is set to NULL then 
to all samples the same fields and records will be added.
}
