% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr_angular_dist}
\alias{tr_angular_dist}
\title{Computes angular distance between two vectors}
\usage{
tr_angular_dist(a, b, w = 1, neg = FALSE)
}
\description{
Computes angular distance between two vectors
}
\keyword{internal}
