% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr_angular_sim}
\alias{tr_angular_sim}
\title{Computes angular similarity between two vectors}
\usage{
tr_angular_sim(a, b, w = 1)
}
\description{
Computes angular similarity between two vectors
}
\keyword{internal}
