% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr_cosine_dist}
\alias{tr_cosine_dist}
\title{Computes cosine distance between two vectors}
\usage{
tr_cosine_dist(a, b, w = 1)
}
\description{
Computes cosine distance between two vectors
}
\keyword{internal}
