% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr_cosine_sim}
\alias{tr_cosine_sim}
\title{Computes cosine similarity between two vectors}
\usage{
tr_cosine_sim(a, b, w = 1)
}
\description{
Computes cosine similarity between two vectors
}
\keyword{internal}
