% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{tr_crop}
\alias{tr_crop}
\title{Cropping a trace}
\usage{
tr_crop(x, crop_to, new_obj = TRUE)
}
\arguments{
\item{x}{an object of class tracer}

\item{crop_to}{sets the limits defining a cropping segment, 
if it is a single positive number,  will be treated as an upper limit 
starting from zero.}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data}
}
\description{
Trims the profile to the set RT range
}
