% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tr_euc2sim}
\alias{tr_euc2sim}
\title{Converts euclidian distance into similarity}
\usage{
tr_euc2sim(a, b, w = 1, lamb = 1, simple = FALSE)
}
\description{
Converts euclidian distance into similarity
}
\keyword{internal}
