% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_reprocess}
\alias{tr_reprocess}
\title{Reprocessing data}
\usage{
tr_reprocess(x, ref = NULL, level = NULL, skip = FALSE)
}
\arguments{
\item{x}{an object of class tracer to reprocess}

\item{ref}{an object of class tracer whose workflow will be used to reprocess data. 
If omitted, a workflow of the object x is going to be used.}

\item{level}{an integer or a vector of integers to specify on which steps 
to perform reprocessing. If a single value is supplied the reprocessing 
will be done for whole chain from the very beginning up to this step included.}

\item{skip}{logical. If TRUE every step will be treated as a separate processing step.}
}
\description{
allows to reprocess Raw data using an existing object as a reference or on its own workflow
}
