% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{tr_resample}
\alias{tr_resample}
\title{Resamples traces}
\usage{
tr_resample(x, pts, new_obj = FALSE)
}
\arguments{
\item{x}{object of class tracer}

\item{pts}{an integer setting the number of points to re-sample}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data}
}
\description{
Re-sample data to new retention times 
 using spline or linear interpolation, using 'prospectr::resample()'
}
