% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_workflow}
\alias{tr_workflow}
\title{Creates a workflow object}
\usage{
tr_workflow(flow)
}
\arguments{
\item{flow}{an expression representing a chain of sequential processing steps}
}
\description{
with workflows you can automate routines  applying the same processing chain 
of operations to a different objects
}
