% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{trace_info}
\alias{trace_info}
\title{Gets trace info}
\usage{
trace_info(x, common = FALSE, extra = TRUE, force_raw = FALSE, ...)
}
\arguments{
\item{x}{an object of class tracer}

\item{common}{logical. If TRUE returns only common field across all samples}

\item{extra}{logical. If TRUE computes and returns additional metrics 
based on meta data}

\item{force_raw}{logical if TRUE returns info for 
the Unprocessed data}

\item{...}{an argument `frac` to pass to `expand_meta_data` function}
}
\value{
a list with descriptive parameters of traces
}
\description{
Provides a compact table with descriptive information
about traces
}
