% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_edge.R
\name{trace_nodes}
\alias{trace_nodes}
\alias{trace_edges}
\title{Accessors for a trace graph}
\usage{
trace_nodes(graph, type = NULL)

trace_edges(graph, rel = NULL)
}
\arguments{
\item{graph}{A \code{trace_graph}.}

\item{type}{Optional node type(s) to filter by.}

\item{rel}{Optional relationship type(s) to filter by.}
}
\value{
A tibble of nodes / edges.
}
\description{
Accessors for a trace graph
}
