% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{trace_sim}
\alias{trace_sim}
\title{Angular Similarity}
\usage{
trace_sim(
  x,
  lab = NULL,
  pw = 0,
  force_raw = FALSE,
  fun = c("max", "mean", "min")
)
}
\arguments{
\item{x}{object of class tracer}

\item{pw}{a numeric, a weighing coefficient for signal intensities
default is 0.}
}
\description{
pair-wisw angular similarity
}
