% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphTracer.R
\name{TracerGraph}
\alias{TracerGraph}
\alias{gtTraces}
\title{Tracer Graph in R}
\usage{
gtTraces(
  mcmc.log,
  burn.in,
  trace.id = c("posterior"),
  drop.1st.row = TRUE,
  line.or.point = 1,
  point.size = 1,
  point.alpha = 1,
  line.alpha = 0.8,
  title = "Trace",
  x.lab = NULL,
  y.lab = NULL,
  legend.title = "parameters",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mcmc.log}{The data frame from MCMC log file whose
column names are parameters and row names are the number
of states at each sample.}

\item{trace.id}{A vector of name(s) of trace(s) to plot.
Default to "posterior".}

\item{drop.1st.row}{If TRUE, also the default, then remove
the 1st row of MCMC log.}

\item{line.or.point}{If 1, the default, then display the line only;
if 2, then only points; if 3, then both the line and points.}

\item{point.size}{The size of points. Default to 1.}

\item{...}{Other arguments passed to \code{\link{readFile}}.}
}
\description{
The visualization functions to understand the Bayesian MCMC result.
}
\details{
\code{gtTraces} returns a \code{\link{gtable}} object to plot
the trace given MCMC log data frame,
whose column names are parameters and row names are the number
of states at each sample.
}
\examples{
mcmc.log <- readMCMCLog("data/star.beast.log")
burn.in.stats <- getBurnIn(rownames(mcmc.log))
burn.in.stats
names(mcmc.log)
gt <- gtTraces(mcmc.log, burn.in.stats$burn.in)
plot(gt)

gt <- gtTraces(mcmc.log, burn.in.stats$burn.in, trace.id=c("TreeHeight.Species", "TreeHeight.t:tree_0_0", "TreeHeight.t:tree_0_1"))

}
\keyword{Tracer}
