% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeStats.R
\name{TreeStats}
\alias{TreeStats}
\alias{readTrees}
\alias{analyseTreeStats}
\title{Summarising tree statistics from the posterior}
\usage{
readTrees(file, burn.in = 0.1, min.trees = 10)

analyseTreeStats(tree.stats, id = c("total.br.len", "tree.height"))
}
\arguments{
\item{file}{The file containing BEAST trees in the nexus format.
The tree should look like "tree STATE_7500 = (1:7.260,(3:3.298 ... ))",
where "tree STATE_7500" is the ID and "7500" is the state of MCMC.}

\item{burn.in}{the proportion of samples treated as burn-in stage from the MCMC.
The calculation of burnin given the proportion is same to Tracer.}

\item{min.trees}{it will throw error, if the number of trees in the file
is smaller than this number, default to 10.}

\item{tree.stats}{The data frame.}
}
\description{
Summarise tree statistics from the posterior trees logged by
BEAST 2 \url{http://www.beast2.org}, such as the total branch length,
tree height, etc.
}
\details{
\code{readTrees} return a data frame of  the posterior trees logged in *.trees file.

\code{analyseTreeStats} get stats from data frame returned by
\code{readTrees}.
\code{id} is set to analyse the default selected columns.
}
\examples{
tre.sta.df <- readTrees("data/RSV2long.trees")

tre.sta <- analyseTreeStats(tre.sta.df)

}
\keyword{TreeStats}
