% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{add_field}
\alias{add_field}
\title{Add a new custom field}
\usage{
add_field(x, new_field, init = NA)
}
\arguments{
\item{x}{an object of class tracer}

\item{new_field}{a string, i.e. a name of the new field}

\item{init}{a value to fill a new field with}
}
\value{
object of type tracer with a new field added into the meta data
}
\description{
adding a field to the meta data
}
