% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{copy_trace}
\alias{copy_trace}
\title{Copy an item of tracer object}
\usage{
copy_trace(x, what = NULL)
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{see `plt_gg` for details}
}
\value{
object of type tracer with selected traces.
}
\description{
copy selected traces from the tracer object
}
