% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_downloadModule.R
\name{createDownloadModuleUI}
\alias{createDownloadModuleUI}
\title{Dynamically add ui elements for downloadButton}
\usage{
createDownloadModuleUI(output, private_key)
}
\arguments{
\item{output}{shiny session output object}

\item{private_key}{Private key used for the conditional logic to add a signed JSON
download button or not}
}
\description{
Dynamically add ui elements for downloadButton
}
