% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{del_trace}
\alias{del_trace}
\title{Deletes selected traces}
\usage{
del_trace(x, what)
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{a smart row selector, see `plt_gg` for details.}
}
\value{
object of type tracer with a new field added into the meta data
}
\description{
removes all selected traces
}
