% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{del_trace}
\alias{del_trace}
\title{Add a new custom field}
\usage{
del_trace(x, what)
}
\arguments{
\item{x}{an object of class tracer}

\item{new_field}{a string, i.e. a name of the new field}
}
\value{
object of type tracer with a new field added into the meta data
}
\description{
adding a field to the meta data
}
