% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{del_trace}
\alias{del_trace}
\title{Delete an item field}
\usage{
del_trace(x, what, keep_history = FALSE)
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{either numeric indexes or vector string of UID. 
If some Numeric indexes out of range an error will be thrown. 
All Unmatched UID's will be ignored with warning.}
}
\value{
object of type tracer with selected traces removed.
}
\description{
delete selected traces from the trace object
}
