% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_downloadModule.R
\name{downloadModuleServer}
\alias{downloadModuleServer}
\title{Download JSON Module Server}
\usage{
downloadModuleServer(id, graph)
}
\arguments{
\item{id}{Module ID}

\item{graph}{reactive graph object to be converted to JSON}
}
\value{
None
}
\description{
Download JSON Module Server
}
