% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_downloadModule.R
\name{downloadModuleUI}
\alias{downloadModuleUI}
\title{Download JSON Module UI}
\usage{
downloadModuleUI(id, label)
}
\arguments{
\item{id}{Module ID}

\item{label}{Button label}
}
\value{
Shiny UI elements
}
\description{
Download JSON Module UI
}
