% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{expand_meta_data}
\alias{expand_meta_data}
\title{Computes some descriptive parameters of a trace}
\usage{
expand_meta_data(lst, fac = 1e+05)
}
\description{
Computes some descriptive parameters of a trace
}
\keyword{.data}
\keyword{importFrom}
\keyword{internal}
\keyword{rlang}
