% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{file_scan}
\alias{file_scan}
\title{Tries different separators and determines a file Template}
\usage{
file_scan(fls, delim = NULL)
}
\description{
Tries different separators and determines a file Template
}
\keyword{internal}
