% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{gen_uid}
\alias{gen_uid}
\title{An ID generator that produces an ID code composed of
lower case letters and 10 digits with prefix 'w'}
\usage{
gen_uid(len)
}
\description{
An ID generator that produces an ID code composed of
lower case letters and 10 digits with prefix 'w'
}
