% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{gen_uid_pool}
\alias{gen_uid_pool}
\title{ID pool generator. Generates a pool of unique ID's}
\usage{
gen_uid_pool(n, len = 6, pool = NULL)
}
\description{
ID pool generator. Generates a pool of unique ID's
}
