% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{gen_uid_pool}
\alias{gen_uid_pool}
\title{ID pool generator}
\usage{
gen_uid_pool(n, len = 6, pool = NULL)
}
\arguments{
\item{n}{a number of UID to generate}

\item{len}{the length or the number of symbols in UID}

\item{pool}{a vector filed with unique ID's - optional.}
}
\description{
Generates a pool of unique ID's
}
