% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_score.R
\name{group_score}
\alias{group_score}
\title{Compute the tracing score of a group of contacts}
\usage{
group_score(x, R, lambda, w)
}
\arguments{
\item{x}{A list of integer vectors indicating dates of onset of the cases
causing exposure. Each item of the list corresponds to a different
individual.}

\item{R}{The average effective reproduction number, i.e. the average number
of secondar cases seeded by an infected individual.}

\item{lambda}{The average number of contacts (infectious or not) reported by
cases.}

\item{w}{The probability mass function of the serial interval, i.e. the delay
from primary to secondary symptom onset. We recommend using the
\code{distcrete} package to generate discretized distributions (see
details).}
}
\value{
A function with one single argument 't', which is the date for which
    the score should be computed.
}
\description{
This function compute the tracing score of a group of contacts, i.e. a set of
individual who have been exposed to one or several cases whose dates of
onsets are known. The tracing score is defined as the expected number of
individuals that will show symptoms for the first time on day 't'.
}
\details{
See the \code{distcrete} package for generating discretized
    distributions at: \url{http://github.com/reconhub/distcrete}.
}
\examples{
if (require(distcrete)) {
 ## generate serial interval distribution
 SI <- distcrete("gamma", 1L, w = 0, 10, 0.65)$d
 plot(SI, type="h", xlim=c(0,50), xlab = "Days")
 title("Serial interval distribution")

 ## get tracing score function for:
 ## - exposure to cases with onsets: 1, 10, 25
 ## - R = 2.1
 ## - lambda = 3.5
 f <- contact_score(c(1,10,25), R=2.1, lambda=3.5, SI)

 ## score for various days
 f(0) # day 0
 f(10) # day 10
 f(10:20)

 ## plot score for various days
 plot(f, type = "h", xlim = c(0,60), col = pal1(100),
      xlab = "Date", ylab = "P(new symptoms)")
 title("Contact score over time")
}
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}

