% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{load_trace}
\alias{load_trace}
\title{LOREM IPSUM}
\usage{
load_trace(path_dir = NULL, fls = NULL, pattern = NULL, uid_len = 6L)
}
\arguments{
\item{path_dir}{a string specifying path to a directory with trace files to load}

\item{fls}{a vector of strings as an alternative to provide file path, overrides
'path_dir'}

\item{pattern}{allows select files by custom regex pattern. NOTE it overrides
selection by file extension pattern}

\item{uid_len}{an integer that specifies the length of unique ID that will
indexing rows.}
}
\value{
an object of class tracer which is a list containing:
 * DATA - a list of two list: RAW and PROCESSED data
 * LOG - a data.frame that contains info about the loading status of files
 * HISTORY - a data.frame with ordered records of processing steps.
 May absent if data is unprocessed.
 * META - a list of meta-data stored separately for each file loaded.
}
\description{
reading in trace files exported from Empower or Chromeleon software.
}
