% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{merge_trace}
\alias{merge_trace}
\title{Merge objects}
\usage{
merge_trace(a, b, what = NULL)
}
\arguments{
\item{a, b}{an objects of class tracer}

\item{what}{an optional argument either numeric indexes or vector string of UID, 
that allows selecting traces from object 'b' to be merged with object 'a'.}
}
\value{
object of type tracer.
}
\description{
merge two trace objects into one
}
\details{
If both arguments 'what'  and 'rehash' are NULL than all identical UIDs found 
in objects 'a' and 'b' will be ignored and not copied into object a. If
argument 'rehash' is TRUE all data will be merged, but all UID's will be recomputed
to ensure uniqueness.
}
