% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{new_trace}
\alias{new_trace}
\title{object constructor}
\usage{
new_trace(x, meta = NULL, use_column = NULL, len = 6)
}
\arguments{
\item{x}{a data.frame or a list of data.frames}

\item{len}{s numeric that defines UID length}

\item{use_names}{if TRUE the names of list items wil be used to fill 
SampleName fieled in the meta data table}
}
\description{
create an obect of type tracer
}
\details{
an input data.frame must contain two columns of type numeric. To construct
an object with multiple traces, a list of data.frames must be provided.
}
