% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{new_trace}
\alias{new_trace}
\title{object constructor}
\usage{
new_trace(x, meta = NULL, use_columns = NULL, use_names = FALSE, len = 6)
}
\arguments{
\item{x}{a data.frame or a list of data.frames}

\item{meta}{meta data}

\item{use_columns}{a vector of length 2}

\item{use_names}{if TRUE than the names of  input list x will be used as ID's}

\item{len}{s numeric that defines UID length}
}
\description{
create an obect of type tracer
}
\details{
an input data.frame must contain two columns of type numeric. To construct
an object with multiple traces, a list of data.frames must be provided.
}
