% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{parse_chromeleon}
\alias{parse_chromeleon}
\title{Chromeleon parser, imports .csv and .txt files}
\usage{
parse_chromeleon(fls, sep)
}
\description{
Chromeleon parser, imports .csv and .txt files
}
\keyword{.data}
\keyword{importFrom}
\keyword{internal}
\keyword{rlang}
