% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_trace_proc.R
\name{parse_empower}
\alias{parse_empower}
\title{Empower parser, imports .csv, .txt, .arw files}
\usage{
parse_empower(fls, skip, sep)
}
\description{
Empower parser, imports .csv, .txt, .arw files
}
\keyword{.data}
\keyword{importFrom}
\keyword{internal}
\keyword{rlang}
