% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{plt_gg}
\alias{plt_gg}
\title{Plot traces with ggplot2}
\usage{
plt_gg(
  x,
  what = NULL,
  stacked = TRUE,
  force_raw = FALSE,
  x_lab = "Retention Time [min]",
  facet_lab = "SampleName",
  xlim = NULL,
  ylim = c(-0.05, 1),
  gr_col = "SampleName",
  breaks = 5,
  minor_breaks = 1,
  expd = 2
)
}
\arguments{
\item{what}{either numeric indexes or string of UID. If some Numeric indexes out of range
an error will be thrown. Unmatched UID's will be ignored with warning.}

\item{force_raw}{If TRUE will plot RAW data}

\item{x_lab}{a string passing x-axis name to the plot}

\item{facet_lab}{trace grouping factors, default available options are names of
a table returning by trace_info.}

\item{xlim, ylim}{numeric vectors defining limits for x and y axis respectively}

\item{breaks}{an integer controlling major breaks on the x-axis}

\item{minor_breaks}{an integer controlling minor breaks on the x-axis}

\item{expd}{expands plot beyond limits set on x-axis}

\item{lst}{a list of data to plot, assumed to originate from a trace object}
}
\description{
plots traces using ggplot2 graphics
}
