% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{plt_gg}
\alias{plt_gg}
\title{Plot traces with ggplot2}
\usage{
plt_gg(
  x,
  what = NULL,
  stacked = TRUE,
  force_raw = FALSE,
  x_lab = "Retention Time [min]",
  facet_lab = "ID",
  xlim = NULL,
  ylim = NULL,
  gr_col = "ID",
  breaks = 5,
  minor_breaks = 1,
  expd = 2
)
}
\arguments{
\item{x}{object of class tracer}

\item{what}{either numeric indexes or string of UID. If some Numeric indexes are 
out of range, an error will be thrown. Unmatched UID's will be ignored with warning.}

\item{stacked}{a logical, sets plotting mode to stacked or overlaid}

\item{force_raw}{If TRUE will plot RAW data}

\item{x_lab}{a string passing x-axis name to the plot}

\item{facet_lab}{trace grouping factors, default available options are names of
a table returning by trace_info.}

\item{xlim, ylim}{numeric vectors defining limits for x and y axis respectively}

\item{gr_col}{name of a veritable to be used to colorize groups}

\item{breaks}{an integer controlling major breaks on the x-axis}

\item{minor_breaks}{an integer controlling minor breaks on the x-axis}

\item{expd}{expands plot beyond limits set on x-axis}
}
\description{
plots traces using ggplot2 graphics
}
