% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{plt_tracer}
\alias{plt_tracer}
\title{Plot traces}
\usage{
plt_tracer(
  x,
  x_lab = "Retention Time [min]",
  force_raw = FALSE,
  facet_lab = "SampleName",
  xlim = c(5, 120),
  ylim = c(-0.05, 1),
  breaks = 5,
  minor_breaks = 1,
  expd = 2
)
}
\arguments{
\item{x}{an object of class tracer}

\item{x_lab}{a string passing x-axis name to the plot}

\item{force_raw}{If TRUE will plot RAW data}

\item{facet_lab}{trace grouping factors, default available options are names of
a table returning by trace_info.}

\item{xlim, ylim}{numeric vectors defining limits for x and y axis respectively}

\item{breaks}{an integer controlling major breaks on the x-axis}

\item{minor_breaks}{an integer controlling minor breaks on the x-axis}

\item{expd}{expands plot beyond limits set on x-axis}
}
\description{
plotting traces - under development -
}
