% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_field_value}
\alias{set_field_value}
\title{Set a new records to a selected field of the meta data}
\usage{
set_field_value(x, what, record, field = c("SampleName", "SRC", "Comments"))
}
\arguments{
\item{x}{an object of class tracer}

\item{what}{either numeric indexes or vector string of UID. 
If some Numeric indexes out of range an error will be thrown. 
All Unmatched UID's will be ignored with warning.}

\item{record}{a vector of entities}

\item{field}{a string representing the name of the field into which 
the record should be placed}
}
\value{
object of type tracer with field records updated
}
\description{
setting new sample name to a trace
}
