% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_align}
\alias{tr_align}
\title{a Trace Aligner}
\usage{
tr_align(x, new_obj = TRUE, ref = 1L, rm_na = TRUE, return_mat = FALSE, ...)
}
\arguments{
\item{x}{an object of class tracer}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data. Also overrides 'return_mat'.}

\item{ref}{an index of a trace that will be used as a reference}

\item{rm_na}{logical, if TRUE will replace all NA's that may be introduced after
ptw alignment algorithm, with the lowest value in the traces.}

\item{return_mat}{if TRUE, instead of a list returns a matrix of aligned data.}

\item{...}{an additional argument to to be passed 'ptw::ptw'.}
}
\value{
either an object of class tracer, or a matrix, or a list of processed data, 
depending on the arguments `return_mat` and `new_obj`, the later overrides `return_mat`.
}
\description{
Aligns traces to a reference provided
}
