% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_baseline}
\alias{tr_baseline}
\title{Besline correction}
\usage{
tr_baseline(x, new_obj = TRUE)
}
\arguments{
\item{x}{an object of class tracer}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data}
}
\description{
flatens besline
}
