% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_compar}
\alias{tr_compar}
\title{Angular Similarity}
\usage{
tr_compar(
  x,
  lab = NULL,
  pw = 0,
  neg = FALSE,
  force_raw = FALSE,
  metric = c("cosim", "cosdist", "angularsim", "angulardist"),
  fun = c("max", "mean", "min")
)
}
\arguments{
\item{x}{object of class tracer}

\item{lab}{name of a meta data column whose values would be used as labels for the output distance 
or similarity matrix}

\item{pw}{a numeric, a weighing coefficient for signal intensities
default is 0.}

\item{neg}{logical, if TRUE allows negative peaks, default is FALSE}

\item{force_raw}{if TRUE RAW data will be compared regardless of the previous 
processing steps taken}

\item{metric}{a string indicating which similarity or distance metric to compute}

\item{fun}{a string defining a function that will be used in getting re-weighting vector.}
}
\description{
pair-wisw angular similarity
}
