% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_resample}
\alias{tr_resample}
\title{Trace re-sampling}
\usage{
tr_resample(x, pts, new_obj = TRUE)
}
\arguments{
\item{x}{an object of class tracer}

\item{pts}{an integer setting the number of points to re-sample}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data}
}
\description{
Re-samples data to new retention time points 
 using spline or linear interpolation, applying 'prospectr::resample'
}
