% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_func.R
\name{tr_rescale}
\alias{tr_rescale}
\title{Re-scale Response of a trace}
\usage{
tr_rescale(x, type = "minmax", new_obj = TRUE, ...)
}
\arguments{
\item{x}{an object of class tracer}

\item{type}{a string defining re-scaling algorithm. Can be 
either minmax with arbitrary limits, or maxnorm that normalizes data to the 
greatest value.}

\item{new_obj}{logical, if TRUE returns a modified object, 
otherwise - processed data}

\item{...}{an additional argument to specify range for minmax_scale rescale.}
}
\description{
Transforms Response to a new designated scale.
}
