% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/trace.R
\name{trace_image}
\alias{trace_image}
\title{Imports}
\usage{
trace_image(stop_window = 20)
}
\arguments{
\item{stop_window}{The number of pixels to create the kill switch.}
}
\value{
A tibble
}
\description{
Click on the image to record the coordinates. When finished click on the
right side of the window. A data frame is returned with the (x, y) coordinates
scaled to (0, 1).
}
\details{
To use \code{trace_image}:
\enumerate{
\item{}{Plot an image or any \code{ggplot}}
\item{}{Run \code{df <- trace_image()}}
\item{}{Click on the desired coordinates in the required sequence}
\item{}{When finished click on the left side of the plot window between 0
and \code{stop_window} pixels on the x-axis}
}

The coordinates are scaled to (0, 1) and returned as a \code{tibble}
}
\examples{
if(interactive()) {
  df <- trace_image()
  inspect_trace(df)
}
}
