% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_meth.R
\name{trace_info}
\alias{trace_info}
\title{Gets trace info}
\usage{
trace_info(x, force_raw = FALSE, ...)
}
\arguments{
\item{x}{object of class tracer}

\item{force_raw}{logical if TRUE returns info for 
the Unprocessed data}

\item{...}{an argument to pass to 'expand_meta_data' function}
}
\value{
a list with descriptive parameters of traces
}
\description{
Provides a compact table with descriptive information
about traces
}
