% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_importHelper.R
\name{updateGraph}
\alias{updateGraph}
\title{Updates the reactive graph element with the uploaded graph}
\usage{
updateGraph(graph, uploadedGraph)
}
\arguments{
\item{graph}{reactive, main graph of the app}

\item{uploadedGraph}{reactive, uploaded graph}
}
\description{
Updates the reactive graph element with the uploaded graph
}
